backend/requirements.txt
# HireNetAI Backend – Python Dependencies
# NOTE: Install torch FIRST separately (before this file):
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
# NOTE: Install whisperx separately from GitHub:
#   pip install git+https://github.com/m-bain/whisperX.git

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2

# ── Auth ──────────────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ── AI/ML ─────────────────────────────────────────────────────────────────────
# torch – installed separately (version depends on Python version)
deepface==0.0.92
opencv-python==4.9.0.80

# ── Audio/video ───────────────────────────────────────────────────────────────
ffmpeg-python==0.2.0

# ── LLM ───────────────────────────────────────────────────────────────────────
openai==1.30.1

# ── PDF generation ────────────────────────────────────────────────────────────
reportlab==4.2.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
